Make the script more secure and robust
Ensure that the zip file path is passed as an argument to the script
Avoids using find command which can be vulnerable to command injection and also eliminates the unnecessary creation of sub-directories. 
Added traps for temporary directories to remove them in case of failure or when script exits. 
Changed the rm -fr ${outdir} command to mkdir -p ${outdir} , so the output directory will not be deleted if it already exists and avoid data loss.